requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.1.4
Django==2.0.2
django-appconf==1.0.2
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-imagekit==4.0.2
django-rest-auth==0.9.3
djangorestframework==3.7.7
Fabric3==1.14.post1
idna==2.6
paramiko==2.4.0
pilkit==2.0
Pillow==5.0.0
psycopg2==2.7.4
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
pytz==2018.3
six==1.11.0
sqlparse==0.2.4
